+++ Source/Fluxor/DependencyInjection/DependencyScanners/EffectClassessDiscovery.cs
@@ -7,14 +7,15 @@ namespace Fluxor.DependencyInjection.DependencyScanners
 {
 	internal static class EffectClassessDiscovery
 	{
-		internal static IEnumerable<DiscoveredEffectClass> DiscoverEffectClasses(
+		internal static DiscoveredEffectClass[] DiscoverEffectClasses(
 			IServiceCollection serviceCollection, IEnumerable<Type> allCandidateTypes)
 		{
-			DiscoveredEffectClass[] discoveredEffectInfos = allCandidateTypes
-				.Where(t => typeof(IEffect).IsAssignableFrom(t))
-				.Where(t => t != typeof(EffectWrapper<>))
-				.Select(t => new DiscoveredEffectClass(implementingType: t))
-				.ToArray();
+			DiscoveredEffectClass[] discoveredEffectInfos =
+				allCandidateTypes
+					.Where(t => typeof(IEffect).IsAssignableFrom(t))
+					.Where(t => t != typeof(EffectWrapper<>))
+					.Select(t => new DiscoveredEffectClass(implementingType: t))
+					.ToArray();
 
 			foreach (DiscoveredEffectClass discoveredEffectInfo in discoveredEffectInfos)
 				serviceCollection.AddScoped(discoveredEffectInfo.ImplementingType);