+++ Source/Fluxor/DependencyInjection/DependencyScanners/EffectMethodsDiscovery.cs
@@ -8,30 +8,29 @@ namespace Fluxor.DependencyInjection.DependencyScanners
 {
 	internal static class EffectMethodsDiscovery
 	{
-		internal static IEnumerable<DiscoveredEffectMethod> DiscoverEffectMethods(IServiceCollection serviceCollection,
+		internal static DiscoveredEffectMethod[] DiscoverEffectMethods(IServiceCollection serviceCollection,
 			IEnumerable<MethodInfo> allCandidateMethods)
 		{
-			DiscoveredEffectMethod[] discoveredEffects = allCandidateMethods
-				.Select(m => new
-				{
-					MethodInfo = m,
-					EffectAttribute = m.GetCustomAttribute<EffectMethodAttribute>(false)
-				})
-				.Where(x => x.EffectAttribute != null)
-				.Select(x => new DiscoveredEffectMethod(
-					hostClassType: x.MethodInfo.DeclaringType,
-					methodInfo: x.MethodInfo,
-					actionType: x.MethodInfo.GetParameters()[0].ParameterType))
-				.ToArray();
+			DiscoveredEffectMethod[] discoveredEffects =
+				allCandidateMethods
+					.Select(m =>
+						new
+						{
+							MethodInfo = m,
+							EffectAttribute = m.GetCustomAttribute<EffectMethodAttribute>(false)
+						})
+					.Where(x => x.EffectAttribute != null)
+					.Select(x => new DiscoveredEffectMethod(x.EffectAttribute, x.MethodInfo))
+					.ToArray();
 
-			IEnumerable<Type> hostClassTypes = discoveredEffects
-				.Select(x => x.HostClassType)
-				.Where(t => !t.IsAbstract)
-				.Distinct();
+			IEnumerable<Type> hostClassTypes =
+				discoveredEffects
+					.Select(x => x.HostClassType)
+					.Where(t => !t.IsAbstract)
+					.Distinct();
 
 			foreach (Type hostClassType in hostClassTypes)
-				if (!hostClassType.IsAbstract)
-					serviceCollection.AddScoped(hostClassType);
+				serviceCollection.AddScoped(hostClassType);
 
 			return discoveredEffects;
 		}