+++ Tests/Fluxor.UnitTests/EffectTests/IsolatedTests.cs
@@ -1,8 +1,0 @@
-﻿namespace Fluxor.UnitTests.EffectTests
-{
-	// This is just here to ensure that reflection tests don't try to
-	// add any classes from this namespace
-	public class IsolatedTests : Middleware
-	{
-	}
-}