+++ Tests/Fluxor.UnitTests/EffectTests/ShouldReactToActionTests/SupportFiles/IsolatedTests.cs
@@ -1,0 +1,6 @@
+﻿namespace Fluxor.UnitTests.EffectTests.ShouldReactToActionTests.SupportFiles
+{
+	public class IsolatedTests : Middleware
+	{
+	}
+}