+++ Tests/Fluxor.UnitTests/MiddlewareTests/IsInsideMiddlewareChangeTests/SupportFiles/IsolatedTests.cs
@@ -1,0 +1,6 @@
+﻿namespace Fluxor.UnitTests.MiddlewareTests.IsInsideMiddlewareChangeTests.SupportFiles
+{
+	public class IsolatedTests : Middleware
+	{
+	}
+}