+++ Tests/Fluxor.UnitTests/MiddlewareTests/OnInternalMiddlewareChangeEndingTests/SupportFiles/IsolatedTests.cs
@@ -1,0 +1,6 @@
+﻿namespace Fluxor.UnitTests.MiddlewareTests.OnInternalMiddlewareChangeEndingTests.SupportFiles
+{
+	public class IsolatedTests : Middleware
+	{
+	}
+}