+++ Tests/Fluxor.UnitTests/StoreTests/AddFeatureTests/IsolatedTests.cs
@@ -1,6 +1,0 @@
-﻿namespace Fluxor.UnitTests.StoreTests.AddFeatureTests
-{
-	public class IsolatedTests : Middleware
-	{
-	}
-}