+++ Tests/Fluxor.UnitTests/StoreTests/BeginInternalMiddlewareChangeTests/IsolatedTests.cs
@@ -1,6 +1,0 @@
-﻿namespace Fluxor.UnitTests.StoreTests.BeginInternalMiddlewareChangeTests
-{
-	public class IsolatedTests : Middleware
-	{
-	}
-}