+++ Tests/Fluxor.UnitTests/StoreTests/DispatchTests/SupportFiles/IsolatedTests.cs
@@ -1,0 +1,6 @@
+﻿namespace Fluxor.UnitTests.StoreTests.DispatchTests.SupportFiles
+{
+	public class IsolatedTests : Middleware
+	{
+	}
+}