+++ Tests/Fluxor.UnitTests/StoreTests/UnhandledExceptionTests/IsolatedTests.cs
@@ -1,6 +1,0 @@
-﻿namespace Fluxor.UnitTests.StoreTests.UnhandledExceptionTests
-{
-	public class IsolatedTests : Middleware
-	{
-	}
-}