+++ Tests/Fluxor.UnitTests/StoreTests/UnhandledExceptionTests/SupportFiles/IsolatedTests.cs
@@ -1,0 +1,6 @@
+﻿namespace Fluxor.UnitTests.StoreTests.UnhandledExceptionTests.SupportFiles
+{
+	public class IsolatedTests : Middleware
+	{
+	}
+}